Apply debian/patches/61_replace_dash_with_bash_mbug675185.dpatch
We have carried along this patch as a patch inside our sources
since 2012 (commit cfd4fcb0bc3d).

The validity of this has thus been vetted in production for years
and the review done now did not find otherwise.

A race in dash causes mysqld_safe to occasionally loop infinitely.
Fix by using bash instead.

https://bugs.launchpad.net/ubuntu/+source/mysql-dfsg-5.0/+bug/675185

As this is the last patch, we can also clean away usage of dpatch.