heap: remove NEW_HASH_FUNCTION
This was added in 29fa72526a22 (in 2000) and nothing defines
NEW_HASH_FUNCTION. A low collision is the wrong tradeoff,
it should be focused on throughput.

d405bee0580eb30e53e5391bbb49ef44fa3d73ad shows the existing function
is pretty cpu intensive and needs to be improved as well.