MDEV-16428: Concurrent DML on RocksDB tables makes optimistic parallel replication abort
DRAFT:
Make MyRocks call thd_rpl_deadlock_check(X, Y) whenever a transaction X
is about to wait on a row that's locked by transaction Y.
This allows the SQL layer to check if transaction Y has "ran ahead" of
transaction X and abort it if it is necessary.

The patch includes adding a hook into RocksDB to detect the waits.