Travis-CI: Ensure AWS key management plugin is not built nor packaged
Building this plugin which requires run-time access to network, uses a lot
of disk space and is slow was already partially disabled. This way we
also ensure that on cmake level it never runs even if it out of some
autodetection reason at times thought it could run.

This fixes the error message:
  fatal: unable to access 'https://github.com/awslabs/aws-sdk-cpp.git/':
  Problem with the SSL CA cert (path? access rights?)