MDEV-15564 Avoid table rebuild in ALTER TABLE on collation or charset changes
Patch makes it's possible to in some cases 'increase' CHARSET length without
touching any data. Thus, only metadata is changed. And this has no penalty for
future DDL or DML queries. Additionally it optimizes UNIQUE and PRIMARY keys
dropping and creation if charset was changed for one of its key parts but only
if collate change does not require change of order in a key. If collations
are not 'compatible' operation can not be performed instantly.

btr_index_rec_validate(): relax check to allow REDUNDANT field be shorter than
specified in index definition